# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customer Service Pi",
    "description": "Account management API \u2014 create, update, look up, and manage customer accounts.\n\n## Authentication\nAll endpoints (except `/health`) require an `X-API-Key` header.\n\n## Pagination\n`GET /accounts` returns paginated results. Use `page` and `page_size` query params.\n\n## Correlation IDs\nEvery request and response carries an `X-Correlation-ID` header. Include this value when raising support tickets or searching logs.",
    "contact": {
      "name": "Platform Engineering",
      "email": "platform@example.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/accounts": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "summary": "Create a new account",
        "description": "Create a new customer account.\n\n- **customer_number** must be unique across all accounts.\n- **email** must be unique and is normalised to lowercase on storage.\n- Returns **201 Created** with the full account representation.",
        "operationId": "create_account_accounts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "409": {
            "description": "Customer number or email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation error in request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Search and list accounts",
        "description": "List and search accounts with pagination.\n\n- Use **query** for case-insensitive substring search.\n- Use **status** to filter by account lifecycle state.\n- Results are ordered by **created_at** descending (newest first).",
        "operationId": "list_accounts_accounts_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search across first_name, last_name, email, customer_number",
              "examples": [
                "Smith"
              ],
              "title": "Query"
            },
            "description": "Free-text search across first_name, last_name, email, customer_number"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AccountStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by account status",
              "title": "Status"
            },
            "description": "Filter by account status"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page (max 100)",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of accounts matching the search criteria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAccounts"
                }
              }
            }
          },
          "422": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_id}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get account by ID",
        "description": "Retrieve a single account by its internal UUID.\n\nPrefer `/accounts/by-customer/{customer_number}` for lookups from business systems.",
        "operationId": "get_account_accounts__account_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full account details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Accounts"
        ],
        "summary": "Partially update an account",
        "description": "Partially update an account (PATCH semantics).\n\nOnly fields present in the request body are modified.\nOmitting a field leaves it unchanged \u2014 do not pass `null` to clear optional fields.",
        "operationId": "update_account_accounts__account_id__patch",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "409": {
            "description": "Email already in use by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "No updatable fields provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Accounts"
        ],
        "summary": "Deactivate an account",
        "description": "Soft-delete an account by setting its status to **inactive**.\n\nPhysical deletion is not supported to preserve audit history.\nUse `PATCH /accounts/{id}` with `{\"status\": \"suspended\"}` to suspend instead.",
        "operationId": "deactivate_account_accounts__account_id__delete",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated account (status set to inactive)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/by-customer/{customer_number}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get account by customer number",
        "description": "Look up an account by its business-facing customer number.\n\nCustomer numbers are case-insensitive; `CUST-001` and `cust-001` resolve to the same account.",
        "operationId": "get_account_by_customer_number_accounts_by_customer__customer_number__get",
        "parameters": [
          {
            "name": "customer_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Customer Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full account details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Health",
          "Health"
        ],
        "summary": "Liveness probe",
        "description": "Simple liveness check \u2014 returns 200 if the process is running.\nGKE kubelet calls this; it must not require authentication (exempt in api_key_auth.py).",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Service is alive",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health/dependencies": {
      "get": {
        "tags": [
          "Health",
          "Health"
        ],
        "summary": "Dependency health and circuit breaker state",
        "description": "Reports the connectivity and circuit breaker state for each downstream service.\n\n- **status: ok** \u2014 dependency is reachable\n- **status: error** \u2014 dependency returned an error or is unreachable\n- **circuit_breaker: open** \u2014 breaker has tripped; calls are being short-circuited\n- **circuit_breaker: closed** \u2014 breaker is healthy; calls are flowing normally\n\n**Important:** an open circuit breaker indicates the *downstream* is struggling.\nCheck this endpoint before restarting this service's pods.",
        "operationId": "health_dependencies_health_dependencies_get",
        "responses": {
          "200": {
            "description": "Live status of all downstream dependencies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependenciesResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccountCreate": {
        "properties": {
          "customer_number": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Customer Number",
            "description": "Unique business identifier (alphanumeric and hyphens only)",
            "examples": [
              "CUST-00123"
            ]
          },
          "first_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "description": "Account holder first name",
            "examples": [
              "Jane"
            ]
          },
          "last_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name",
            "description": "Account holder last name",
            "examples": [
              "Smith"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Primary email address (must be unique across accounts)",
            "examples": [
              "jane.smith@example.com"
            ]
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "Contact phone number in any international format",
            "examples": [
              "+1-800-555-0100"
            ]
          }
        },
        "type": "object",
        "required": [
          "customer_number",
          "first_name",
          "last_name",
          "email"
        ],
        "title": "AccountCreate",
        "description": "Payload for POST /accounts."
      },
      "AccountResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Internal UUID \u2014 use customer_number for lookups"
          },
          "customer_number": {
            "type": "string",
            "title": "Customer Number"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_number",
          "first_name",
          "last_name",
          "email",
          "phone",
          "status",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by"
        ],
        "title": "AccountResponse",
        "description": "Full account representation returned by all read/write endpoints."
      },
      "AccountStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "suspended"
        ],
        "title": "AccountStatus"
      },
      "AccountSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "customer_number": {
            "type": "string",
            "title": "Customer Number"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_number",
          "first_name",
          "last_name",
          "email",
          "status",
          "created_at"
        ],
        "title": "AccountSummary",
        "description": "Lightweight projection used in paginated list results."
      },
      "AccountUpdate": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name",
            "examples": [
              "Janet"
            ]
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name",
            "examples": [
              "Smith-Jones"
            ]
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "examples": [
              "janet.smith@example.com"
            ]
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "examples": [
              "+44 20 7946 0958"
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountStatus"
              },
              {
                "type": "null"
              }
            ],
            "description": "Account lifecycle status",
            "examples": [
              "inactive"
            ]
          }
        },
        "type": "object",
        "title": "AccountUpdate",
        "description": "Payload for PATCH /accounts/{id}.\nAll fields optional \u2014 only provided fields are updated (true partial update)."
      },
      "DependenciesResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "dependencies": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DependencyStatus"
            },
            "type": "object",
            "title": "Dependencies"
          }
        },
        "type": "object",
        "required": [
          "status",
          "dependencies"
        ],
        "title": "DependenciesResponse"
      },
      "DependencyStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "circuit_breaker": {
            "type": "string",
            "title": "Circuit Breaker"
          }
        },
        "type": "object",
        "required": [
          "status",
          "circuit_breaker"
        ],
        "title": "DependencyStatus"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable error code",
            "examples": [
              "ACCOUNT_NOT_FOUND"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable error description"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id",
            "description": "X-Correlation-ID from the request \u2014 include this in support tickets"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Standard error response envelope \u2014 never exposes internal stack traces."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse"
      },
      "PaginatedAccounts": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccountSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of matching records"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page (1-based)"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "description": "Number of items per page"
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "Total number of pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PaginatedAccounts",
        "description": "Envelope for paginated list responses."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}